Show an info bar if the sender request MDN
Show an info bar if the sender request a disposition notification
after the user read the message. At this info bar, there is a
button that the user should click to send the message. Notice,
however, that this commit does not implement the bits to actually
send the message.

It is also added in this commit the signals necessary to send the
e-mail with the MDN. Hence, if the user click the `Send
Notification` button, then the following signal is triggered:

    ConversationMessage => ConversationEmail => GearyController

Finally, the function `on_send_disposition_notification` must be
used to send the e-mail with the MDN.