Use <HR> to separate two text parts
E-mails with two or more text parts are shown in Geary as it was
just one, because it appends the text of each part without any
separations. This commit adds a <HR> between the text parts if the
text is HTML, or a '------------------------------' otherwise.

Fixes bug#775532.

It also improve the coding style (fix indentation, remove trailing
spaces, wrap the code to be inside 80 columns, etc.) of the
function `construct_body_from_mime_parts`.